server/requirements.txt
fastapi
uvicorn
python-magic